refactor: extract DTO mapping tests to WatchZoneSummaryDTOTests
Moves toDomain() unit tests into a dedicated test suite to keep
APIWatchZoneRepositoryTests under the 400-line SwiftLint file_length
limit. The loadAll integration test for mixed valid/invalid zones
stays in the repository test suite.
